requirements.txt
flask>=2.0.0
werkzeug>=2.0.0
python-docx>=0.8.10
pypdf2>=2.0.0
scikit-learn>=1.0.0